\docType{data}
\name{dmrs}
\alias{dmrs}
\title{Example bumphunter output}
\usage{
dmrs
}
\description{
The result of running the bump hunting algorithm to find DMRs between
colon and kidney (20 samples each from TCGA), with B = 1000 permutations, and a
cutoff of 0.2 on the Beta values.
}
\keyword{datasets}
